Rename march_hardware_configuration_parser to march_hardware_builder
