+++ march_hardware_builder/yaml/dummy_temperature.yaml
