+++ march_hardware_interface/CMakeLists.txt
@@ -14,7 +14,7 @@ find_package(catkin REQUIRED COMPONENTS
   std_msgs
   controller_manager
   march_hardware
-  march_hardware_configuration_parser
+  march_hardware_builder
 )
 
 catkin_package(
